autoscraper/requirements.txt
# AutoScraper Service Dependencies
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM (MongoDB)
beanie>=1.24.0
motor>=3.3.2
pymongo>=4.6.0

# Redis & Caching
redis>=5.0.1
hiredis>=2.2.3

# Background Tasks
celery>=5.3.4
kombu>=5.3.4

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
playwright>=1.40.0
selenium>=4.15.2
scrapy>=2.11.0

# HTTP Client
httpx>=0.25.2
aiohttp>=3.9.1

# Data Processing
pandas>=2.1.4
numpy>=1.24.4
python-dateutil>=2.8.2
feedparser>=6.0.10

# Text Processing
nltk>=3.8.1
spacy>=3.7.2
textblob>=0.17.1

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
email-validator>=2.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
loguru>=0.7.2

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.1.0

# Security
cryptography>=41.0.7
bcrypt>=4.1.2

# Configuration
pyyaml>=6.0.1
toml>=0.10.2

# Rate Limiting
slowapi>=0.1.9

# Job Scheduling
apscheduler>=3.10.4

# File Processing
openpyxl>=3.1.2

# Network
requests-oauthlib>=1.3.1
urllib3>=2.1.0

# Async Support
anyio>=4.2.0
asyncio-mqtt>=0.16.1

# Data Validation
cerberus>=1.3.5
marshmallow>=3.20.2

# Performance
psutil>=5.9.6
memory-profiler>=0.61.0

# Browser Automation
selenium-wire>=5.1.0
webdriver-manager>=4.0.1

# AI Integration
openai>=1.3.0
anthropics>=0.8.0
tiktoken>=0.5.2